Example of using server routes instead of /api
