+++ nuxt.config.ts
@@ -6,6 +6,7 @@ export default defineNuxtConfig({
     modules: ['@nuxtjs/tailwindcss', '@sidebase/nuxt-auth'],
     auth: {
         origin: process.env.ORIGIN,
+        basePath: '/auth',
         enableGlobalAppMiddleware: true
     }
 })
